Add context
This commit adds all of the necessary logic to support Contexts similar
to what you find in Golang.

A context is an object that you can interact with to decide if you
should exit from an independent async operation. It's especially useful
inside of tokio::task::spawn.

By practice you pass the context in, and then you usually will call
done() as one of the branches of a tokio::select! with the other branch
being a long running future.

This allows us to not only set a timeout for that task,
but to also retain a handle in the parent calling function that can be
used to immediately cancel the context at any time.

Doing so will cause done() to no longer block and return immediately,
causing the internal select() to then exit.