Simplify Context and TaskController APIs
This commit simplifies the APIs for both Context and TaskController by
removing the optional duration argument for their respective new
constructors. Instead I've added separate constructors for constructing
with a timeout. This simplifies call sites by not requiring any
arguments in the case that someone doesn't want a timeout, and reducing
an unnecessary use of an option type in cases when they do want one.